Hij is nu iets soepeler en explosies zijn goed.

+++ src/level/level.h
@@ -76,10 +76,10 @@ class Level
 
 	//Bomb getBomb(int index);
 	//void setBomb(int index, Bomb bomb);
-	void setBomb(int index, uint8_t x, uint8_t y, uint8_t time, uint8_t peep);
+	void setBomb(int index, uint8_t x, uint8_t y, uint32_t time, uint8_t peep);
 	uint8_t getBombX(int index);
 	uint8_t getBombY(int index);
-	uint8_t getBombTime(int index);
+	uint32_t getBombTime(int index);
 	uint8_t getBombPeep(int index);
 
 
@@ -111,7 +111,7 @@ class Level
 
 	uint8_t bombX[2] = {0};
 	uint8_t bombY[2] = {0};
-	uint8_t bombTime[2] = {0};
+	uint32_t bombTime[2] = {0};
 	uint8_t bombPeep[2] = {0};
 
 	// The name of the level
